Introduce additional checks that promote hocon maps to json arrays.
In addition to being a valid integer keys must also form a consecutive 0..N-1 set of numbers (N being size of map).
Previous implementation was treating any map with integer keys as arrays, not forcing them to be consecutive and effectively
prohibiting actual maps of some keys that look like integers to be maps.

The workaround for map

{ "1" : "A" , "5" : "B" }

was to introduce artifical non-integer key .

After this change, such map is no loneger treated like array because it has only elements on indices 1,5 not forming a consecutive range 0,1

The rationale behind this change is that this is hocon-json bridge not pure hocon config. Thus it is typically
used in scenarios where there was a json processing already in place and move to hocon is driven by it features like optional commas, multiline string, etc.
however max compliancy with json is required.
Existing json structures still may relay on arrays and maps and they will fail when maps are unexpectedly converted to arrays.